requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bidict==0.22.0
boto3==1.24.26
botocore==1.27.26
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-jsonfield==1.4.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
multidict==6.0.2
mysql-connector-python==8.0.31
numpy==1.23.1
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.49
Pillow==9.2.0
protobuf==3.20.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.0
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
sentry-sdk==1.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.9
yarl==1.7.2
